Fix POD links, declare non-core prereqs
I noted that one of the POD links to the GitHub repo was incorrect, so I fixed that and dutifully added an author test (which can be run with `make authortest`). Note that checking HTTPS scheme URLs and uninstalled modules via CPAN Meta DB requires the installation of Net::SSLeay and IO::Socket::SSL; I've opened @trwyant/perl-Test-Pod-LinkCheck-Lite#2 to get that resolved.

That led me down a rabbit hole of making sure this could build, test, and install on Perl 5.10 as declared in the Makefile.PL. There were a couple issues:

1. URL links in POD didn't gain support for displaying text instead of the bare URL until Perl 5.12. So I adjusted the POD to use unlinked text with the links adjacent.
2. Some core Perl module dependencies either weren't core at that time (JSON::PP) or lacked features (File::Path's `make_path` function, version's `declare` method). So I added explicit version requirements for both that contain the required functions.

Also, just for completeness' sake, I adjusted the Makefile.PL to skip the META_MERGE Makefile argument on older versions of ExtUtils::MakeMaker.